Moved all interactions with api to single file
